+++ simgui-window.json
@@ -5,11 +5,11 @@
   "MainWindow": {
     "GLOBAL": {
       "fps": "120",
-      "height": "1467",
+      "height": "1055",
       "maximized": "0",
       "style": "0",
       "userScale": "6",
-      "width": "3171",
+      "width": "1924",
       "xpos": "386",
       "ypos": "463"
     }
@@ -51,17 +51,17 @@
     "###Addressable LEDs": {
       "Collapsed": "0",
       "Pos": "1521,48",
-      "Size": "339,886"
+      "Size": "273,78"
     },
     "###FMS": {
       "Collapsed": "0",
       "Pos": "17,1003",
-      "Size": "653,272"
+      "Size": "546,236"
     },
     "###Joysticks": {
       "Collapsed": "0",
       "Pos": "879,979",
-      "Size": "2056,472"
+      "Size": "1696,337"
     },
     "###Keyboard 0 Settings": {
       "Collapsed": "0",
@@ -80,7 +80,7 @@
     },
     "###Other Devices": {
       "Collapsed": "0",
-      "Pos": "2213,28",
+      "Pos": "1905,28",
       "Size": "691,868"
     },
     "###RoboRIO": {
@@ -91,12 +91,12 @@
     "###System Joysticks": {
       "Collapsed": "0",
       "Pos": "10,524",
-      "Size": "475,470"
+      "Size": "394,398"
     },
     "###Timing": {
       "Collapsed": "0",
       "Pos": "11,252",
-      "Size": "337,276"
+      "Size": "281,240"
     },
     "Debug##Default": {
       "Collapsed": "0",
@@ -106,7 +106,7 @@
     "Robot State": {
       "Collapsed": "0",
       "Pos": "16,46",
-      "Size": "213,204"
+      "Size": "178,174"
     }
   }
 }
